requirements.txt
tensorflow-gpu==1.15.0
scipy==1.1.0
